DEV(cache_critical_dns): sort resolved SRV targets by priority
The priority field in an SRV RR indicates a preferential order at which the
underlying targets should be utilised. We need to attempt to resolve
healthy services in order of priority, where 0 is highest.

It may seem weird that the SRV target addresses are resolved in reverse
order of priority, but this is necessary; when all targets are healthy
we pick the _most recently resolved_ target first. In this case, the
highest priority target, if healthy, should be resolved after all lower
priority targets.

Before sorting, we still ensure that we only consider targets with a
priority within the given threshold as previously implemented.

Prior to this commit, we relied on whatever order the
dnsclient.getresources method returned. As it turns out, this assumption
is incorrect. The order returned is likely whatever order the system
resolver received DNS responses in, which may not be ordered according
to the spec.